added rating driver, rider feature and swagger api

+++ src/main/java/in/yash/UberApplication/services/DriverService.java
@@ -16,7 +16,7 @@ public interface DriverService {
 
     RideDto endRide(Long rideId);
 
-    RiderDto rateRider(Long rideId, Integer rating);
+    RiderDto rateRider(Long rideId, Double rating);
 
     DriverDto getMyProfile();
